requirements.txt
celery==4.2.1
coreapi==2.3.3
Django==3.0.2
django-cors-headers==3.5.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-celery-beat==2.0.0
django-ninja==0.14.0
django-redis==4.12.1
factory-boy==2.12.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
redis==3.5.3

Pillow~=8.0.1
django-model-utils~=4.0.0
whitenoise==5.2.0
xmltodict==0.12.0